requirements.txt
# Core
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0

# Scraping
praw==7.7.1
newsapi-python==0.2.7
feedparser==6.0.11

# Sentiment Analysis
vaderSentiment==3.3.2

# Stock & Market Data (real-time price, fundamentals, history)
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4

# Optional: FinBERT (uncomment for better accuracy, requires ~500MB)
# transformers==4.42.4
# torch==2.3.1
